requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.5
psycopg2-binary==2.9.6
# psycopg2==2.9.9
django-filter==23.2
python-dotenv==1.0.0
celery==5.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
# django-celery-monitor==1.1.2
djangorestframework_simplejwt==5.2.2
django-cryptography==1.1
django-cryptography-django5==2.2
celery-once==3.0.1
sqlalchemy==2.0.13
# выключил проверку на HTTPS
django-cors-headers==4.0.0
graypy==2.1.0
django-dbbackup==4.0.2
rapidfuzz==3.0.0
docx2txt==0.8
python-ldap==3.4.4
django-auth-ldap==4.5.0
django-remote-auth-ldap==1.0.0
django-axes==6.1.0
# gssapi==1.8.3
python-decouple==3.8
pyodbc==5.0.1
pypsrp==0.8.1
paramiko==3.3.1
PyYAML==6.0.1
atlassian-python-api==3.39.0
beautifulsoup4==4.12.2
requests==2.31.0
graypy==2.1.0
Jinja2==3.1.2
django-auditlog==2.3.0
emoji==2.2.0
django-widget-tweaks==1.5.0
pandas==2.2.2
openpyxl==3.1.2
tenacity==9.0.0
django-jazzmin==3.0.0
retrying==1.3.4